+++ projects/igniteui-angular/src/lib/combo/combo.component.spec.ts
@@ -658,7 +658,6 @@ describe('igxCombo', () => {
             combo.comboInput = {
                 value: '',
             } as any;
-            combo.filteringOptions.filterable = true;
             const matchSpy = spyOn<any>(combo, 'checkMatch').and.callThrough();
             spyOn(combo.searchInputUpdate, 'emit');
 
@@ -681,11 +680,6 @@ describe('igxCombo', () => {
             expect(matchSpy).toHaveBeenCalledTimes(3);
             expect(combo.searchInputUpdate.emit).toHaveBeenCalledTimes(2);
             expect(combo.searchInputUpdate.emit).toHaveBeenCalledWith(args);
-
-            combo.filteringOptions.filterable = false;
-            combo.handleInputChange();
-            expect(matchSpy).toHaveBeenCalledTimes(4);
-            expect(combo.searchInputUpdate.emit).toHaveBeenCalledTimes(2);
         });
         it('should be able to cancel searchInputUpdate', () => {
             combo = new IgxComboComponent(elementRef, mockCdr, mockSelection as any, mockComboService,
@@ -693,7 +687,6 @@ describe('igxCombo', () => {
             spyOn(mockIconService, 'addSvgIconFromText').and.returnValue(null);
             combo.ngOnInit();
             combo.data = data;
-            combo.filteringOptions.filterable = true;
             combo.searchInputUpdate.subscribe((e) => {
                 e.cancel = true;
             });
@@ -890,7 +883,6 @@ describe('igxCombo', () => {
                 expect(combo.itemHeight).toEqual(32);
                 expect(combo.placeholder).toEqual('Location');
                 expect(combo.searchPlaceholder).toEqual('Enter a Search Term');
-                expect(combo.filteringOptions.filterable).toEqual(true);
                 expect(combo.allowCustomValues).toEqual(false);
                 expect(combo.cssClass).toEqual(CSS_CLASS_COMBO);
                 expect(combo.type).toEqual('box');
@@ -1084,14 +1076,6 @@ describe('igxCombo', () => {
                 const focusedItem_2 = dropdownItems[5];
                 expect(focusedItem_2.classList.contains(CSS_CLASS_FOCUSED)).toBeTruthy();
                 expect(focusedItem_1.classList.contains(CSS_CLASS_FOCUSED)).toBeFalsy();
-            });
-            it(`should not render search input if both 'allowCustomValues' and 'filterable' are false`, () => {
-                combo.allowCustomValues = false;
-                combo.filteringOptions.filterable = false;
-                expect(combo.displaySearchInput).toBeFalsy();
-                combo.toggle();
-                fixture.detectChanges();
-                expect(combo.searchInput).toBeFalsy();
             });
             it('should focus search input', fakeAsync(() => {
                 combo.toggle();
@@ -2594,7 +2578,7 @@ describe('igxCombo', () => {
                 combo.filterFunction = comboIgnoreDiacriticsFilter;
                 combo.displayKey = null;
                 combo.valueKey = null;
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: undefined };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: undefined };
                 combo.data = ['José', 'Óscar', 'Ángel', 'Germán', 'Niño', 'México', 'Méxícó', 'Mexico', 'Köln', 'München'];
                 combo.toggle();
                 fixture.detectChanges();
@@ -2865,50 +2849,7 @@ describe('igxCombo', () => {
                 tick();
                 fixture.detectChanges();
                 expect(combo.dropdown.items.length).toEqual(0); // No items are available because of filtering
-
-                combo.close(); // Filter is cleared on close
-                tick();
-                fixture.detectChanges();
-                combo.filteringOptions.filterable = false; // Filtering is disabled
-                fixture.detectChanges();
-                combo.open(); // All items are visible since filtering is disabled
-                tick();
-                fixture.detectChanges();
-                expect(combo.dropdown.items.length).toBeGreaterThan(0); // All items are visible since filtering is disabled
-
-                combo.searchValue = 'Not-available item';
-                combo.handleInputChange();
-                fixture.detectChanges();
-                expect(combo.dropdown.items.length).toBeGreaterThan(0); // All items are visible since filtering is disabled
-
-                combo.close(); // Filter is cleared on close
-                tick();
-                fixture.detectChanges();
-                combo.filteringOptions.filterable = true; // Filtering is re-enabled
-                fixture.detectChanges();
-                combo.open(); // Filter is cleared on open
-                tick();
-                fixture.detectChanges();
-                expect(combo.dropdown.items.length).toBeGreaterThan(0);
             }));
-            it(`should properly display "Add Item" button when filtering is off`, () => {
-                combo.allowCustomValues = true;
-                combo.filteringOptions.filterable = false;
-                fixture.detectChanges();
-                expect(combo.isAddButtonVisible()).toEqual(false);
-
-                combo.toggle();
-                fixture.detectChanges();
-                expect(combo.collapsed).toEqual(false);
-                const searchInput = fixture.debugElement.query(By.css(CSS_CLASS_SEARCHINPUT));
-                UIInteractions.triggerInputEvent(searchInput, combo.data[2].field);
-                fixture.detectChanges();
-                expect(combo.isAddButtonVisible()).toEqual(false);
-
-                UIInteractions.triggerInputEvent(searchInput, combo.searchValue.substring(0, 2));
-                fixture.detectChanges();
-                expect(combo.isAddButtonVisible()).toEqual(true);
-            });
             it('should be able to toggle search case sensitivity', () => {
                 combo.showSearchCaseIcon = true;
                 fixture.detectChanges();
@@ -2949,7 +2890,7 @@ describe('igxCombo', () => {
                 combo.close();
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: combo.groupKey };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: combo.groupKey };
                 combo.filterFunction = (collection: any[], searchValue: any, filteringOptions: IComboFilteringOptions): any[] => {
                     if (!collection) return [];
                     if (!searchValue) return collection;
@@ -2987,7 +2928,7 @@ describe('igxCombo', () => {
                 combo.close();
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: combo.groupKey };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: combo.groupKey };
                 combo.filterFunction = (collection: any[], searchValue: any, filteringOptions: IComboFilteringOptions): any[] => {
                     if (!collection) return [];
                     if (!searchValue) return collection;
@@ -3024,7 +2965,7 @@ describe('igxCombo', () => {
                 combo.close();
                 tick();
                 fixture.detectChanges();
-                combo.filteringOptions = { caseSensitive: false, filterable: true, filteringKey: combo.groupKey };
+                combo.filteringOptions = { caseSensitive: false, filteringKey: combo.groupKey };
                 combo.open();
                 tick();
                 fixture.detectChanges();
@@ -3039,10 +2980,6 @@ describe('igxCombo', () => {
                 combo.handleInputChange();
                 fixture.detectChanges();
                 expect(combo.dropdown.items.length).toEqual(0);
-
-                combo.filteringOptions = Object.assign({}, combo.filteringOptions, { filterable: false });
-                fixture.detectChanges();
-                expect(combo.dropdown.items.length).toBeGreaterThan(0);
             }));
         });
         describe('Form control tests: ', () => {
@@ -3368,7 +3305,7 @@ describe('igxCombo', () => {
 @Component({
     template: `
     <igx-combo #combo [placeholder]="'Location'" [data]='items'
-        [filterable]='true' [valueKey]="'field'" [groupKey]="'region'" [width]="'400px'"
+        [valueKey]="'field'" [groupKey]="'region'" [width]="'400px'"
         (selectionChanging)="selectionChanging($event)" [style.--ig-size]="'var(--ig-size-' + size + ')'">
         <ng-template igxComboItem let-display let-key="valueKey">
             <div class="state-card--simple">
@@ -3451,7 +3388,7 @@ class IgxComboSampleComponent {
         </p>
         <p>
             <igx-combo #comboReactive formControlName="townCombo"
-                class="input-container" [filterable]="true" placeholder="Location(s)"
+                class="input-container" placeholder="Location(s)"
                 [data]="items" [displayKey]="'field'" [groupKey]="'region'">
                 <label igxLabel>Town</label>
             </igx-combo>
@@ -3525,7 +3462,7 @@ class IgxComboFormComponent {
     <form #form="ngForm">
         <igx-combo #testCombo class="input-container" [placeholder]="'Locations'"
             name="anyName" required [(ngModel)]="values"
-            [data]="items" [filterable]="filterableFlag"
+            [data]="items"
             [displayKey]="'field'" [valueKey]="'field'"
             [groupKey]="'field' ? 'region' : ''" [width]="'100%'">
             <label igxLabel>Combo Label</label>
@@ -3625,8 +3562,7 @@ export class IgxComboBindingTestComponent {
     <div class="comboContainer" [style.width]="'500px'">
         <igx-combo #combo placeholder="Location(s)"
             [data]="citiesData"
-            [allowCustomValues]="true"
-            [filterable]="true">
+            [allowCustomValues]="true">
         </igx-combo>
     </div>
     `,