+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes for each version of this project will be documented in this file.
 
+## 18.1.0
+### General
+- `IgxCombo` and  `IgxSimpleComboComponent`
+    - **Breaking Change** `filterable` is removed from the `filteringOptions`.
+
 ## 18.0.0
 ### New Features
 - `IgxCombo`, `IgxSimpleCombo`: